requirements.txt
aiogram==3.13.1
aiomysql==0.2.0
black==25.1.0
flake8==7.1.1
flake8-import-order==0.18.2
flake8-quotes==3.4.0
pep8-naming==0.14.1
python-dotenv==1.0.1
selenium==4.31.0
SQLAlchemy==2.0.36
greenlet==3.1.1
